% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{loadH5}
\alias{loadH5}
\title{loadH5}
\usage{
loadH5(
  FileName,
  assay = "RNA",
  SeuratVersion = checkSeuratVersion(),
  calData = TRUE,
  calScale = FALSE,
  calFeatures = FALSE,
  datatype = "Seurat"
)
}
\arguments{
\item{readType}{}
}
\description{
loadH5
}
