% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{readH5}
\alias{readH5}
\title{readH5}
\usage{
readH5(
  FileName,
  assay = "RNA",
  SeuratVersion = checkSeuratVersion(),
  calData = TRUE,
  calScale = FALSE,
  calFeatures = FALSE,
  readType = "Seurat"
)
}
\arguments{
\item{readType}{}
}
\description{
readH5
}
