% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{saveH5}
\alias{saveH5}
\title{saveH5}
\usage{
saveH5(
  FileName,
  data,
  assay = "RNA",
  save_graph = TRUE,
  SeuratVersion = checkSeuratVersion(),
  split_save = TRUE,
  max_cells_per_subset = 5000
)
}
\arguments{
\item{max_cells_per_subset}{}
}
\description{
saveH5
}
